Rehash fix bug

+++ ConcatStringTree.cpp
@@ -36,8 +36,9 @@ CSTNode::CSTNode(int _leftLength=0, int _length = 0, string _data = "", CSTNode 
 }
 
 // CST SECTION
-CST::CST(CSTNode *_root = nullptr,  bool _isTemporary = false, bool _isShallowNorDeep = false): 
+CST::CST(CSTNode *_root = nullptr, int _numOfNode = 0 ,bool _isTemporary = false, bool _isShallowNorDeep = false): 
     root(_root),
+    numOfNode(_numOfNode),
     isTemporary(_isTemporary),
     isShallowNorDeep(_isShallowNorDeep)
 {
@@ -45,7 +46,8 @@ CST::CST(CSTNode *_root = nullptr,  bool _isTemporary = false, bool _isShallowNo
 
 CST::CST(const CST&& otherS) {
     this->root = nullptr;
-    this->isShallowNorDeep = false;
+    this->numOfNode = otherS.numOfNode;
+    this->isShallowNorDeep = otherS.isShallowNorDeep;
     this->isTemporary = false;
     if (otherS.isShallowNorDeep) {//is return from concat so we shallow copy
         (this->root) = (otherS.root);
@@ -88,7 +90,8 @@ CST::CST(const char * _s){
     
     CSTNode* cstnode = new CSTNode(0, s.size(), s, nullptr, nullptr);
     PTNode * ptnode  = new PTNode(cstnode,0,0, nullptr, nullptr);
-    ptnode->newID();
+    ptnode->newID(1);
+    numOfNode = 1;
     cstnode->parent.insert(cstnode->parent.root, ptnode);
     
     root = cstnode;
@@ -251,8 +254,8 @@ string CST::toString() const{
 ConcatStringTree CST::concat(const ConcatStringTree & otherS) const{
 
     CSTNode *cstnode = new CSTNode(this->length(), this->root->length+ otherS.length(),"", this->root, otherS.root);
-    CST result(cstnode, true, true);
-
+    CST result(cstnode, this->numOfNode+ otherS.numOfNode , true, true);
+    
     //parentTree
     result.createParentAndChildAncestor();
     return (CST&&)result;
@@ -272,7 +275,8 @@ ConcatStringTree CST::subString(int from, int to) const{
         int from;
         int to;
         int current;
-        
+        int numOfNode;
+
         static CSTNode* executeFunc(CSTNode * root, TempStruct &result, CSTNode *left, CSTNode *right){
             if(root->left == nullptr && root->right == nullptr){//is string node
                 string tpStr = "";
@@ -286,6 +290,7 @@ ConcatStringTree CST::subString(int from, int to) const{
                 result.current+= root->length;
 
                 if(tpStr.size()!= 0){
+                    result.numOfNode++;
                     CSTNode *cstnode = new CSTNode(0,tpStr.size(),tpStr);
                     return cstnode;
                 }
@@ -293,7 +298,7 @@ ConcatStringTree CST::subString(int from, int to) const{
             }
             else{//non-string node                
                 if(left == nullptr && right == nullptr) return nullptr;
-
+                result.numOfNode++;
                 int leftLength = 0, length = 0 ;
                 if(left) {
                     leftLength = left->length;
@@ -310,13 +315,10 @@ ConcatStringTree CST::subString(int from, int to) const{
 
     };
     
-    TempStruct obj{from , to , 0};
+    TempStruct obj{from , to , 0,0};
     CSTNode * tempRoot = postorder(this->root, obj, TempStruct::executeFunc); 
 
-    CST result;
-    result.root = tempRoot;
-    result.isTemporary = true;
-    result.isShallowNorDeep =false;
+    CST result(tempRoot,obj.numOfNode, true, false);
 
     //cout<<result.toStringPreOrder()<<endl;
     return (CST&&) result;
@@ -357,10 +359,7 @@ ConcatStringTree CST::reverse() const{
     TempStruct obj;
     CSTNode * tempRoot = postorder(this->root, obj, TempStruct::executeFunc); 
 
-    CST result;
-    result.root = tempRoot;
-    result.isTemporary = true;
-    result.isShallowNorDeep =false;
+    CST result(tempRoot, this->numOfNode, true, false);
 
     //cout<<result.toStringPreOrder()<<endl;
     return (CST&&) result;
@@ -416,10 +415,9 @@ void CST::createParentAndChildAncestor() const{
     };
     
     PTNode *ptnode = new PTNode(this->root, 0,0,nullptr,nullptr);
-    ptnode->newID();
-
+    if(this->isShallowNorDeep == 1) ptnode->newID(1);
+    else ptnode->newID(this->numOfNode);
     TempStruct obj{ptnode};
-
     this->root->parent.insert(this->root->parent.root, ptnode);
     
     CSTNode * left = this->root->left , *right = this->root->right;
@@ -453,13 +451,19 @@ PTNode::PTNode( CSTNode* _data = nullptr, int _id = 0, int _height = 0, PTNode *
 {
 }
 
-void PTNode::newID(){
+void PTNode::newID(int increase = 1){
     maxId++;
+    
+    id = maxId; 
+    this->data->myId = id;
     if(maxId>= 10000000){
         throw overflow_error("Id is overflow!");
     }
-    id = maxId; 
-    this->data->myId = id;
+
+    maxId+= increase-1;
+    if(maxId>= 10000000){
+        throw overflow_error("Id is overflow!");
+    }
 }
 
 
@@ -519,50 +523,69 @@ LSH::~LSH(){
     }
 }
 
-ll LSH::hashFunc(string s, int m){
+ll LSH::hashFunc(string s, ll m){
     ll result = 0, currP = 1;
     FOR(i,0,s.size()){
-        result+= ((ll)s[i])*currP;
-        currP *= config.p; 
+        result+= ((ll)s[i]%m)*(currP%m)%m;
+        currP = ((currP%m)* (config.p%m))%m; 
     }
     result = result% m;
     return result;
 }
 
-int LSH::probingFunc(LitString **&currLitHash,string s, int m,bool acceptNull){
-    int index = hashFunc(s, m), hashedValue = index;
-    ll maxI =(ll)m*(ll)m*(ll)m; 
+ll LSH::probingFunc(LitString **&currLitHash,string s, ll m,bool acceptNull){
+    ll index = hashFunc(s, m), hashedValue = index;
+    ll maxI = m; 
+    
     FOR(i,0,maxI){
-        index = ((int)( (double)hashedValue + config.c1*(double)i+ config.c2*(double)(i*i)+m))%m; 
-        
-        if(currLitHash[index] == NULL && acceptNull){
-            break;
+        ll c1 = config.c1;
+        ll c2 = config.c2;
+
+
+        ll decimalPart = (config.c1- c1)*i + (config.c2 - c2)*((i*i));
+        c1 = (i*c1)%m;
+        c2 = (i*i*c2)%m;
+        index = ( (hashedValue+ decimalPart) + (c1 + c2)%m)%m ; 
+
+        //index = (ll)((double) hashedValue+ config.c1*(double)i + config.c2*(double)(i*i)+m)%m;
+        //cout<<s<<" i = "<<i<<" index= "<<index<<" "<<(ll)((double) hashedValue+ config.c1*(double)i + config.c2*(double)(i*i)+m)<<endl;
+        if(currLitHash[index] == nullptr && acceptNull){
+            return index;
         }
         
         if(currLitHash[index] == nullptr) continue;
         if(currLitHash[index]->s == s){
-            break;
+
+            return index;
         }
     }
-    return index;
+
+    return -1;
 }
 
 void LSH::rehash(){
-    int nSize = config.alpha* (double)config.initSize;
+    ll nSize = (config.alpha* (double)config.initSize);
     LitString **nHash = new LitString*[nSize];
-    
+
+    bool nLastInserted = true;
     FOR(i, 0, nSize){
         nHash[i] = nullptr;
     }
     FOR(i,0,config.initSize){
         LitString *current = litHash[i];
         if(current == nullptr){
-            nHash[i] = nullptr;
             continue;
         }
         LitString *nLitString = new LitString(current->data, current->s, current->reference);
-        int nIndex = probingFunc(nHash,current->s, nSize, true);
+        ll nIndex = probingFunc(nHash,current->s, nSize, true);
+        if(nIndex == -1) {
+            throw runtime_error("No possible slot");
+        }
         nHash[nIndex] = nLitString;
+        if(i == lastInsertedIndex && nLastInserted) {
+            lastInsertedIndex = nIndex;
+            nLastInserted = false;
+        }
         delete current;
     }
 
@@ -577,7 +600,11 @@ CSTNode* LSH::insert(string s){
         throw runtime_error("No possible slot");
     }
     
-    int litIndex = probingFunc(litHash,s , config.initSize, true);
+    ll litIndex = probingFunc(litHash,s , config.initSize, true);
+    
+    if(litIndex == -1) {
+        throw runtime_error("No possible slot");
+    }
     CSTNode *cstnode = nullptr;
     if(litHash[litIndex] == nullptr){
         cstnode = new CSTNode(0, (int)s.size(), s, nullptr, nullptr);
@@ -590,16 +617,20 @@ CSTNode* LSH::insert(string s){
         cstnode = litHash[litIndex]->data;
     }
 
-    if((double)numOfElement/(double)config.initSize > config.lambda) rehash();
-    //cout << "litHash[litIndex] " << litHash[litIndex]->s << " = " << litHash[litIndex]->reference << endl;
+
+    //cout<<"index = "<<litIndex << ", litHash[litIndex] = " << litHash[litIndex]->s << ", ref = " << litHash[litIndex]->reference << endl;
+    if((double)numOfElement/(double)config.initSize > config.lambda) {
+        //cout<<"DEBUG: Before rehash : "<<this->toString()<<endl;
+        rehash();
+    }
         
     return cstnode;
 }
 
 void LSH::remove(string s) {
-    int index = probingFunc(litHash, s, config.initSize,false);
-    if (litHash[index] == NULL) {
-        throw runtime_error("NO STRING FOUND");
+    ll index = probingFunc(litHash, s, config.initSize,false);
+    if ( index == -1 || litHash[index] == NULL) {
+        throw runtime_error("No possible slot");
         return;
     }
     //cout << "litHash[index] " << litHash[index]->s << " = " << litHash[index]->reference << endl;
@@ -610,6 +641,7 @@ void LSH::remove(string s) {
         delete litHash[index];
         litHash[index] = nullptr;
         numOfElement--;
+        if(index == lastInsertedIndex) lastInsertedIndex = -1;
     }
     
 
@@ -646,19 +678,15 @@ RCST::RCST(const char * _s, LitStringHash * _litStringHash){
     litStringHash = _litStringHash;
     CSTNode *cstnode = litStringHash->insert(s);
     
-    PTNode * ptnode  = new PTNode(cstnode,0,0, nullptr, nullptr);
-    ptnode->newID();
-    cstnode->parent.insert(cstnode->parent.root, ptnode);
-    
     root = cstnode;
-
+    numOfNode = 1;
     isShallowNorDeep = false;
     isTemporary = false;
 
 }
 
-RCST::RCST(CSTNode * _root = nullptr, LitStringHash * _litStringHash = nullptr, bool _isTemporay = false, bool _isShallowNorCopy = false):
-   CST(_root, _isTemporay, _isShallowNorCopy), litStringHash(_litStringHash)
+RCST::RCST(CSTNode * _root = nullptr, LitStringHash * _litStringHash = nullptr, int _numOfNode = 0,bool _isTemporay = false, bool _isShallowNorCopy = false):
+   CST(_root, _numOfNode, _isTemporay, _isShallowNorCopy), litStringHash(_litStringHash)
 {
 
 }
@@ -666,7 +694,8 @@ RCST::RCST(CSTNode * _root = nullptr, LitStringHash * _litStringHash = nullptr, 
 RCST::RCST(RCST && otherS){
     this->root = nullptr;
     this->litStringHash = otherS.litStringHash;
-    this->isShallowNorDeep = false;
+    this->numOfNode = otherS.numOfNode;
+    this->isShallowNorDeep = otherS.isShallowNorDeep;
     this->isTemporary = false;
     if (otherS.isShallowNorDeep) {//is return from concat so we shallow copy
         (this->root) = (otherS.root);
@@ -732,16 +761,13 @@ RCST::~RCST(){
         static void ancestorFunc(CSTNode * root, TempStruct &result){
             if (root->ancestor.findPTNode(root->ancestor.root, result.target)) {
                 root->ancestor.deleteNode(root->ancestor.root, result.target);
-
             }
         }  
         static CSTNode* deleteFunc(CSTNode * root, TempStruct &result, CSTNode* left, CSTNode *right){
             if (root->left == nullptr && root->right == nullptr && root->data != "") {
                 result.litStringHash->remove(root->data);
-
             }
             else if (root->ancestor.size() + root->parent.size() == 0){//is non-string node
-                
                 root->left = nullptr;
                 root->right = nullptr;
                 delete root;
@@ -750,8 +776,15 @@ RCST::~RCST(){
         }  
     };
     
+    if(root->left == nullptr && root->right == nullptr && root->data != ""){
+        litStringHash->remove(root->data);
+        isTemporary = true;
+        return;
+    }
+
     int  deletingId = root->myId;
-    root->parent.deleteNode(root->parent.root, deletingId);
+    if (root->parent.findPTNode(root->parent.root,deletingId))
+        root->parent.deleteNode(root->parent.root,deletingId);
     
     TempStruct obj{deletingId, litStringHash};
     CSTNode * left = this->root->left , *right = this->root->right;
@@ -784,7 +817,7 @@ RCST::~RCST(){
 ReducedConcatStringTree RCST::concat(const ReducedConcatStringTree& otherS) const {
 
     CSTNode* cstnode = new CSTNode(this->length(), this->root->length + otherS.length(), "", this->root, otherS.root);
-    RCST result(cstnode, this->litStringHash, true, true);
+    RCST result(cstnode, this->litStringHash, this->numOfNode+otherS.numOfNode,true, true);
 
     //parentTree
     result.createParentAndChildAncestor();
@@ -804,6 +837,8 @@ ReducedConcatStringTree RCST::subString(int from, int to) const{
         int from;
         int to;
         int current;
+
+        int numOfNode;
         
         static CSTNode* executeFunc(CSTNode * root, TempStruct &result, CSTNode *left, CSTNode *right){
             if(root->left == nullptr && root->right == nullptr){//is string node
@@ -818,7 +853,7 @@ ReducedConcatStringTree RCST::subString(int from, int to) const{
                 result.current+= root->length;
 
                 if(tpStr.size()!= 0){
-
+                    result.numOfNode++;
                     CSTNode *cstnode = new CSTNode(0,tpStr.size(),tpStr);
                     return cstnode;
                 }
@@ -827,6 +862,7 @@ ReducedConcatStringTree RCST::subString(int from, int to) const{
             else{//non-string node                
                 if(left == nullptr && right == nullptr) return nullptr;
 
+                result.numOfNode++;
                 int leftLength = 0, length = 0 ;
                 if(left) {
                     leftLength = left->length;
@@ -843,10 +879,10 @@ ReducedConcatStringTree RCST::subString(int from, int to) const{
 
     };
     
-    TempStruct obj{from , to , 0};
+    TempStruct obj{from , to , 0, 0};
     CSTNode * tempRoot = postorder(this->root, obj, TempStruct::executeFunc); 
 
-    RCST result(tempRoot, litStringHash, true, false);
+    RCST result(tempRoot, litStringHash,obj.numOfNode, true, false);
 
     //cout<<result.toStringPreOrder()<<endl;
     return (RCST&&) result;
@@ -887,8 +923,54 @@ ReducedConcatStringTree RCST::reverse() const{
     TempStruct obj;
     CSTNode * tempRoot = postorder(this->root, obj, TempStruct::executeFunc); 
 
-    RCST result(tempRoot, litStringHash, true, false);
+    RCST result(tempRoot, litStringHash, this->numOfNode,true, false);
     //cout<<result.toStringPreOrder()<<endl;
     return (RCST&&) result;
 
+}
+
+
+
+void RCST::createParentAndChildAncestor() const{
+    
+    struct TempStruct{   
+        PTNode *target;
+        static void ancestorFunc(CSTNode * root, TempStruct &result){
+            if(!root->ancestor.findPTNode(root->ancestor.root,result.target->id)){
+                PTNode *ptnode = new PTNode(result.target->data, result.target->id,0,nullptr,nullptr);
+                root->ancestor.insert(root->ancestor.root,ptnode);
+            } 
+        }  
+    };
+    
+    PTNode *ptnode = new PTNode(this->root, 0,0,nullptr,nullptr);
+    if(this->isShallowNorDeep == 1) ptnode->newID(1);
+    else ptnode->newID(this->numOfNode);
+
+    TempStruct obj{ptnode};
+
+    this->root->parent.insert(this->root->parent.root, ptnode);
+    
+    CSTNode * left = this->root->left , *right = this->root->right;
+    if(left){
+        if(!left->parent.findPTNode(left->parent.root, ptnode->id)) {
+            
+            PTNode *leftNode = new PTNode(ptnode->data, ptnode->id,0,nullptr,nullptr);
+            left->parent.insert(left->parent.root, leftNode);
+        }
+
+        preorder(left->left, obj, TempStruct::ancestorFunc);
+        preorder(left->right, obj, TempStruct::ancestorFunc);
+    }
+
+    if(right){
+        
+        if(!right->parent.findPTNode(right->parent.root,ptnode->id)){
+            PTNode *rightNode = new PTNode(ptnode->data, ptnode->id,0,nullptr,nullptr);
+            right->parent.insert(right->parent.root, rightNode);
+        } 
+        
+        preorder(right->left, obj, TempStruct::ancestorFunc);
+        preorder(right->right, obj, TempStruct::ancestorFunc);
+    }
 }
